move out openai_provider related code in llm_openai
