+++ llm_tool/models/__init__.py
@@ -1,6 +1,6 @@
 from . import llm_thread
 from . import llm_tool
-from . import llm_provider_openai
+from . import llm_provider
 from . import llm_tool_record_retriever
 from . import llm_tool_server_action
 from . import llm_model
