+++ llm_tool/models/llm_provider.py
@@ -1,0 +1,14 @@
+import json
+import logging
+
+from odoo import models
+
+_logger = logging.getLogger(__name__)
+
+
+class LLMProvider(models.Model):
+    _inherit = "llm.provider"
+
+    def format_tools_for_provider(self, tools):
+        """Format tools for the specific provider"""
+        return self._dispatch("format_tools", tools)
